nfa: improve build time for interleaved NFA
The main issue with this NFA is that the build time can be quite long,
due to the search for the right place to interleave each state.
A quite naive implementation would start the search at the beginning of
the table for each state, leading to O(n^2) algorithm.

As can be seen in the benchmarks from the previous commit, the current
implementation is still quite slow, with more that 3 seconds to build
an NFA with 65k states.

To improve it, take a page from the original implementation at
https://github.com/mischasan/aho-corasick (implementation that actually
differs from the one described in the http://goo.gl/lE6zG paper): keep
a indicate search index for each offset, and advance it by an arbitrary
heuristic for the offset that yielded the max search index.

This probably can seem quite arbitrary and not that useful, but the
improvements are dramatic. The patterns of the bench from the previous
commit changes like this:

before: build time: 5.33s, size: 3.48MB
after: build time: 74.6ms, size: 3.49MB

The size of the automaton barely increased, while the build time went
down significantly, coming closer to the contiguous NFA build times.