feat: optimize interleave NFA with max_match/special ids
Force some ordering of the states when interleaving them to be able to
use the max_special_id and max_match_id values for the checks done
on every step of the automaton.

This is a bit tricky to make it work, but it does improve the
performance of the algorithm by a significant margin.